+++ packages/cli-plugin-check/src/utils.ts
@@ -7,3 +7,11 @@ export const transformToRelative = (baseDir: string, targetDir: string) => {
     return targetDir;
   }
 };
+
+export const getMainVersion = version => {
+  version = (version || '').split('.')[0];
+  if (version === 'latest' || version === 'beta') {
+    return '3';
+  }
+  return version.replace(/[^\d]/g, '') || '3';
+};